requirements.txt
fastapi==0.129.0
uvicorn==0.40.0
groq==1.0.0
python-dotenv==1.2.1